fix: Ghostlike PyEpics connection timeout problem.
It is discovered that importing ophyd.ophydobj will
trigger this error. It only occurs in the slave
process generated by pty.fork(), not inside the
master thread, which results in the debugging process
being excruciating and confusing.
Personally, I guess the ultimate reason of this subtle
bug is inside some non-standard feature of python
abused by ophyd, like weakref, etc. Anyway, using these
feature just because 'it is there' can not constitute
a proper reason for using them, while it is also safe
and friendly to keep things simple and sound. I hope
you the reader will bear this in mind.